+++ crates/length-log-core/src/ports/error.rs
